expose some prototype transforms utils
