+++ test/test_prototype_transforms_utils.py
@@ -6,8 +6,8 @@ import torch
 from prototype_common_utils import make_bounding_box, make_detection_mask, make_image
 
 from torchvision.prototype import features
-from torchvision.prototype.transforms._utils import has_all, has_any
 from torchvision.prototype.transforms.functional import to_image_pil
+from torchvision.prototype.transforms.utils import has_all, has_any
 
 
 IMAGE = make_image(color_space=features.ColorSpace.RGB)
